move code into src folder and rename bin as q. update setup.py accordingly
