+++ src/q/.qrc
