+++ src/q/tests/test_suite.py
@@ -41,7 +41,7 @@ import re
 import collections
 
 sys.path.append(os.path.join(os.path.abspath(os.path.dirname(sys.argv[0])),'..','bin'))
-from bin.q import QTextAsData, QOutput, QOutputPrinter, QInputParams, DataStream, Sqlite3DB
+from q.q import QTextAsData, QOutput, QOutputPrinter, QInputParams, DataStream, Sqlite3DB
 
 # q uses this encoding as the default output encoding. Some of the tests use it in order to 
 # make sure that the output is correctly encoded
